+++ src/parser/mod.rs
@@ -4,7 +4,4 @@ mod members;
 mod other;
 mod types;
 
-pub(crate) use other::{
-    multispace_or_comment0, multispace_or_comment1, parse_arguments, parse_dictionary_members,
-    parse_ext_attrs, parse_identifier, parse_members, parse_quoted_string,
-};
+pub(crate) use other::*;
